+++ llvm/lib/Target/RISCV/MCTargetDesc/RISCVFixupKinds.h
@@ -71,21 +71,17 @@ getRelocPairForSize(unsigned Size) {
   default:
     llvm_unreachable("unsupported fixup size");
   case 1:
-    return std::make_pair(
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_ADD8),
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_SUB8));
+    return std::make_pair(FirstLiteralRelocationKind + ELF::R_RISCV_ADD8,
+                          FirstLiteralRelocationKind + ELF::R_RISCV_SUB8);
   case 2:
-    return std::make_pair(
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_ADD16),
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_SUB16));
+    return std::make_pair(FirstLiteralRelocationKind + ELF::R_RISCV_ADD16,
+                          FirstLiteralRelocationKind + ELF::R_RISCV_SUB16);
   case 4:
-    return std::make_pair(
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_ADD32),
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_SUB32));
+    return std::make_pair(FirstLiteralRelocationKind + ELF::R_RISCV_ADD32,
+                          FirstLiteralRelocationKind + ELF::R_RISCV_SUB32);
   case 8:
-    return std::make_pair(
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_ADD64),
-        MCFixupKind(FirstLiteralRelocationKind + ELF::R_RISCV_SUB64));
+    return std::make_pair(FirstLiteralRelocationKind + ELF::R_RISCV_ADD64,
+                          FirstLiteralRelocationKind + ELF::R_RISCV_SUB64);
   }
 }