+++ mlir/lib/Target/LLVMIR/Dialect/NVVM/LLVMIRToNVVMTranslation.cpp
@@ -15,7 +15,6 @@
 #include "mlir/Target/LLVMIR/ModuleImport.h"
 
 #include "llvm/IR/ConstantRange.h"
-#include "llvm/IR/IntrinsicsNVPTX.h"
 
 using namespace mlir;
 using namespace mlir::NVVM;